sentiment/requirements.txt
aws-psycopg2
requests
transformers